+++ packages/insured-bridge-relayer/src/RelayerHelpers.ts
@@ -21,7 +21,6 @@ export async function approveL1Tokens(
   const bridgeAdmin = new web3.eth.Contract(getAbi("BridgeAdminInterface"), bridgeAdminAddress);
 
   for (const whitelistedL1Token of whitelistedRelayL1Tokens) {
-    console.log("whitelistedL1Token", whitelistedL1Token);
     const bridgePool = (await bridgeAdmin.methods.whitelistedTokens(whitelistedL1Token).call()).bridgePool;
     if (bridgePool === ZERO_ADDRESS) throw new Error("whitelistedRelayL1Tokens contains not-whitelisted token");
     const approvalTx = await setAllowance(web3, gasEstimator, account, bridgePool, whitelistedL1Token);
@@ -36,7 +35,7 @@ export async function approveL1Tokens(
   }
 }
 
-// Returns the L2 Deposit box address for a given bridgeAdmin
+// Returns the L2 Deposit box address for a given bridgeAdmin on L1.
 export async function getL2DepositBoxAddress(web3: Web3, chainId: number, bridgeAdminAddress: string): Promise<string> {
   const depositContracts = await new web3.eth.Contract(getAbi("BridgeAdminInterface"), bridgeAdminAddress).methods
     .depositContracts(chainId)