+++ packages/serverless-orchestration/src/ServerlessSpoke.js
@@ -18,6 +18,7 @@ const exec = require("child_process").exec;
 const { Logger, delay } = require("@uma/financial-templates-lib");
 let logger;
 
+const waitForLogsS = 5;
 spoke.post("/", async (req, res) => {
   try {
     logger.debug({
@@ -59,7 +60,7 @@ spoke.post("/", async (req, res) => {
       childProcessIdentifier: _getChildProcessIdentifier(req),
       execResponse,
     });
-    await delay(2); // Wait a few seconds to be sure the the winston logs are processed upstream.
+    await delay(waitForLogsS); // Wait a few seconds to be sure the the winston logs are processed upstream.
     res.status(200).send({
       message: "Process exited with no error",
       childProcessIdentifier: _getChildProcessIdentifier(req),
@@ -75,7 +76,7 @@ spoke.post("/", async (req, res) => {
       jsonBody: req.body,
       execResponse: execResponse instanceof Error ? execResponse.message : execResponse,
     });
-    await delay(2); // Wait a few seconds to be sure the the winston logs are processed upstream.
+    await delay(waitForLogsS); // Wait a few seconds to be sure the the winston logs are processed upstream.
     res.status(500).send({
       message: "Process exited with error",
       childProcessIdentifier: _getChildProcessIdentifier(req),