+++ packages/contracts-frontend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-frontend@0.1.2...@uma/contracts-frontend@0.1.3) (2021-10-01)
+
+**Note:** Version bump only for package @uma/contracts-frontend
+
+## [0.1.2](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-frontend@0.1.1...@uma/contracts-frontend@0.1.2) (2021-10-01)
+
+**Note:** Version bump only for package @uma/contracts-frontend
+
 ## [0.1.1](https://github.com/UMAprotocol/protocol/compare/@uma/contracts-frontend@0.1.0...@uma/contracts-frontend@0.1.1) (2021-09-28)
 
 ### Bug Fixes